requirements.txt
Django==1.4
PyYAML==3.10
South==0.7.5
amqplib==1.0.2
anyjson==0.3.1
argh==0.15.1
celery==2.5.3
certifi==0.0.8
chardet==1.0.1
distribute==0.6.27
django-celery==2.5.5
django-contact-form==0.3.1
django-fields==0.2.0
django-picklefield==0.2.1
django-social-auth==0.6.9
django-supervisor==0.2.7
gunicorn==0.14.3
httplib2==0.7.4
kombu==2.1.8
-e git://github.com/brad/mailsnake.git@503dd6a8ba1426edf13fff13be2ff9c0354f35ae#egg=mailsnake
meld3==0.6.8
oauth2==1.5.211
oauthlib==0.1.3
pathtools==0.1.2
pyasn1==0.1.3
pycrypto==2.6
python-dateutil==1.5
python-openid==2.2.5
python-twitter==0.8.2
redis==2.4.13
requests==0.13.1
rsa==3.0.1
simplejson==2.5.2
supervisor==3.0a12
tweepy==1.9
watchdog==0.6.0
wsgiref==0.1.2
